move serch input and btn addNew from TodoList to TodosModule & update styles in some files
